requirements_spark.txt
py4j==0.10.9.7
# pyspark==3.5.0
python-dotenv==1.0.0
# matplotlib>=3.0
# python3-scipy 
# python3-simpy
pandas>=1.0.5                      #Required for pandas API on Spark and Spark Connect; Optional for Spark SQL
pyarrow>=4.0.0                     #Required for pandas API on Spark and Spark Connect; Optional for Spark SQL
numpy>=1.15                        #Required for pandas API on Spark and MLLib DataFrame-based API; Optional for Spark SQL
grpcio>=1.48,<1.57                 #Required for Spark Connect
grpcio-status>=1.48,<1.57          #Required for Spark Connect
googleapis-common-protos ==1.56.4  #Required for Spark Connect
snowflake-connector-python==3.5.0
